first commit
camdrem
camdrem is flash card program that will use the super memo 2 algorithm.
The program will have users who will have multiple decks of flash cards.
Multiple decks can contain the same card. User Progress and schudule tracking
is included and done at the card level.

In this commit the data layer is done and has been tested.  There
are some stubs classes and methods in the model layer but much work
is left to be done.  column ids used at teh data layer are currently
hidden from the model layer but this may prove to be a mistake and is
currently up for debate.